LPD-0 Refator components that are using withProjects to use useProjects hook
