+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/shared/hoc/index.js
@@ -18,12 +18,10 @@ export {default as withHelpWidget} from './WithHelpWidget';
 export {default as withHistory} from './WithHistory';
 export {default as withIndividual} from './WithIndividual';
 export {default as withInterval} from './WithInterval';
-export {default as WithJoinableProjects} from './WithJoinableProjects';
 export {default as withOnboarding} from './WithOnboarding';
 export {default as withPaginationBar} from './WithPaginationBar';
 export {default as withPermission, withAdminPermission} from './WithPermission';
 export {default as withProject} from './WithProject';
-export {default as withProjects} from './WithProjects';
 export {default as withQuery} from './WithQuery';
 export {default as withQueryPagination} from './WithQueryPagination';
 export {default as withQueryRangeSelectors} from './WithQueryRangeSelectors';
