+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/tsconfig.json
@@ -12,5 +12,7 @@
 		"sourceMap": true,
 		"target": "ES5"
 	},
-	"include": ["./src/main/js/**/*"]
+	"include": [
+		"./src/main/js/**/*"
+	]
 }
